+++ crates/ruma-client-api/src/room.rs
@@ -15,7 +15,7 @@ use crate::PrivOwnedStr;
 
 /// Whether or not a newly created room will be listed in the room directory.
 #[doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/doc/string_enum.md"))]
-#[derive(Clone, Default, PartialEq, Eq, StringEnum)]
+#[derive(Clone, Default, StringEnum)]
 #[ruma_enum(rename_all = "snake_case")]
 #[non_exhaustive]
 pub enum Visibility {