+++ crates/ruma-client-api/src/room/create_room.rs
@@ -194,7 +194,7 @@ pub mod v3 {
 
     /// A convenience parameter for setting a few default state events.
     #[doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/doc/string_enum.md"))]
-    #[derive(Clone, PartialEq, Eq, StringEnum)]
+    #[derive(Clone, StringEnum)]
     #[ruma_enum(rename_all = "snake_case")]
     #[non_exhaustive]
     pub enum RoomPreset {