+++ crates/ruma-client-api/src/uiaa.rs
@@ -233,7 +233,7 @@ impl<'de> Deserialize<'de> for AuthData {
 
 /// The type of an authentication stage.
 #[doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/doc/string_enum.md"))]
-#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, StringEnum)]
+#[derive(Clone, StringEnum)]
 #[non_exhaustive]
 pub enum AuthType {
     /// Password-based authentication (`m.login.password`).