+++ crates/ruma-events/src/key/verification/cancel.rs
@@ -66,7 +66,7 @@ impl KeyVerificationCancelEventContent {
 ///
 /// Custom error codes should use the Java package naming convention.
 #[doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/doc/string_enum.md"))]
-#[derive(Clone, PartialEq, Eq, StringEnum)]
+#[derive(Clone, StringEnum)]
 #[ruma_enum(rename_all = "m.snake_case")]
 #[non_exhaustive]
 pub enum CancelCode {