+++ crates/ruma-events/src/room/member.rs
@@ -278,7 +278,7 @@ impl SyncRoomMemberEvent {
 
 /// The membership state of a user.
 #[doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/doc/string_enum.md"))]
-#[derive(Clone, PartialEq, Eq, StringEnum)]
+#[derive(Clone, StringEnum)]
 #[ruma_enum(rename_all = "lowercase")]
 #[non_exhaustive]
 pub enum MembershipState {