+++ crates/ruma-events/src/room_key_request.rs
@@ -48,7 +48,7 @@ impl ToDeviceRoomKeyRequestEventContent {
 
 /// A new key request or a cancellation of a previous request.
 #[doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/doc/string_enum.md"))]
-#[derive(Clone, PartialEq, Eq, StringEnum)]
+#[derive(Clone, StringEnum)]
 #[ruma_enum(rename_all = "snake_case")]
 #[non_exhaustive]
 pub enum Action {